+++ src/Controller/CartValidateController.php
@@ -20,32 +20,40 @@ class CartValidateController extends AbstractController
     {
         // Get the user object
         $user = $this->getUser();
-        $user_id = $user->getId();
+        if($user){
+            $user_id = $user->getId();
+            if (!empty($cartService->getCart($user_id))){
 
-        $invoice = new Invoice();
-        $invoice->setUser($user);
-        $invoice->setDateTransaction(date_create("now"));
+                $invoice = new Invoice();
+                $invoice->setUser($user);
+                $invoice->setDateTransaction(date_create("now"));
 
-        $total_price = $cartService->getTotalPrice($user_id);
-        $invoice->setTotalPrice($total_price);
+                $total_price = $cartService->getTotalPrice($user_id);
+                $invoice->setTotalPrice($total_price);
 
-        $form = $this->createForm(InvoiceFormType::class, $invoice);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $invoice = $form->getData();
+                $form = $this->createForm(InvoiceFormType::class, $invoice);
+                $form->handleRequest($request);
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $invoice = $form->getData();
 
-            $invoice->setTotalPrice($total_price);
-            dump('totalprice'.$total_price);
-            dump($invoice);
-            $soldUserService->decreaseSold($total_price,$user);
-            $cartService->removeCart($user);
-            
-            $entityManager->persist($invoice);
-            $entityManager->flush();
+                    $invoice->setTotalPrice($total_price);
+                    dump('totalprice'.$total_price);
+                    dump($invoice);
+                    $soldUserService->decreaseSold($total_price,$user);
+                    $cartService->removeCart($user);
+                    
+                    $entityManager->persist($invoice);
+                    $entityManager->flush();
 
-            return $this->redirectToRoute('homepage');
+                    return $this->redirectToRoute('homepage');
+                }
+            }else{
+                $this->addFlash('error','tu ne peux pas valider un panier vide');
+                return $this->redirectToRoute('homepage');
+            }
+        }else{
+            return $this->redirectToRoute('app_login');
         }
-
         return $this->render('cart_validate/index.html.twig', [
             'form' => $form ->createView(),
         ]);
